+++ examples/apps/site/handler/handler.js
@@ -1,8 +1,7 @@
 const fs = require("fs");
 const path = require("path");
 const mime = require("mime-types");
-
-const ssrCache = {};
+const isutf8 = require("isutf8");
 
 const createResponse = ({ type, body, isBase64Encoded, headers }) => {
     return {
@@ -21,26 +20,29 @@ module.exports.handler = async event => {
     let key = event.pathParameters ? event.pathParameters.key : "";
     let type = mime.lookup(key);
     let isBase64Encoded = false;
-    console.log("KEY", key);
+
+    if (key.endsWith("undefined")) {
+        return createResponse({
+            type,
+            body: "",
+            isBase64Encoded,
+            headers: { "Cache-Control": "public, max-age=60" }
+        });
+    }
 
     if (!type) {
         type = "text/html";
-        if (!ssrCache[key]) {
-            const { handler } = require("./ssr");
-            ssrCache[key] = await handler("/" + key);
-        }
+        const { html } = await require("./renderer")("/" + key);
 
         return createResponse({
             type,
-            body: ssrCache[key],
+            body: html,
             isBase64Encoded,
             headers: { "Cache-Control": "public, max-age=60" }
         });
     }
 
     const filePath = path.resolve(key);
-
-    // TODO: check if file should be base64 encoded (binary files)
 
     try {
         let buffer = await new Promise((resolve, reject) => {
@@ -50,6 +52,7 @@ module.exports.handler = async event => {
             });
         });
 
+        isBase64Encoded = !isutf8(buffer);
         const headers = {};
 
         if (key.includes("static")) {