+++ packages/cli/sls/deploy.js
@@ -1,22 +1,78 @@
 const { join, resolve } = require("path");
 const fs = require("fs");
-const { green } = require("chalk");
+const { green, red } = require("chalk");
+const notifier = require("node-notifier");
 const execute = require("./execute");
-const { isApiEnvDeployed } = require("./utils");
+const { isApiEnvDeployed, isAppsEnvDeployed } = require("./utils");
 
-const perks = ["a 🍪", "☕️", "an 🍎", "a 🍺"];
+const perks = ["a 🍪", "☕️", "an 🍎", "a 🍺", "a 🥤"];
+
+const notify = ({ message }) => {
+    notifier.notify({
+        title: "Webiny CLI",
+        message,
+        icon: join(__dirname, "logo.png"),
+        sound: false,
+        wait: true
+    });
+};
 
 module.exports = async inputs => {
     const { what, env } = inputs;
 
+    const webinyJs = resolve("webiny.js");
+    const config = require(webinyJs);
+
     if (what === "apps") {
+        if (typeof env === "undefined") {
+            console.log(
+                `🚨 You must specify the ${green("--env")} to deploy when running ${green(
+                    "deploy-apps"
+                )} command.`
+            );
+            process.exit(1);
+        }
+
         if (env === "local") {
             console.log(
                 `🚨 You can't deploy "local" apps environment as it is reserved for local development.`
             );
             process.exit(1);
         }
-        return await execute(inputs);
+
+        // Check if all apps have the requested environment in .env.json
+        for (let i = 0; i < config.apps.length; i++) {
+            const app = config.apps[i];
+            const appLocation = resolve(app.location);
+            const envJson = require(join(appLocation, ".env.json"));
+            if (!envJson[env]) {
+                console.log(
+                    `🚨 Environment ${green(env)} is missing in ${green(
+                        app.name
+                    )} app. Check your ${green(app.location + "/.env.json")} file.`
+                );
+                process.exit(1);
+            }
+        }
+
+        const isFirstDeploy = !(await isAppsEnvDeployed(env));
+
+        const { output, duration } = await execute(inputs);
+
+        console.log(`\n🎉 Done! Deploy finished in ${green(duration + "s")}.`);
+
+        notify({ message: `Apps deploy completed in ${duration}s.` });
+
+        if (output.cdn) {
+            console.log(`🚀 Your website URL: ${green(output.cdn.url)}`);
+            if (isFirstDeploy) {
+                console.log(
+                    `⏳ Please note that CDN distribution takes some time to propagate, so expect this URL to become accessible in ~10 minutes.`
+                );
+            }
+        }
+
+        return;
     }
 
     if (what === "api") {
@@ -30,16 +86,19 @@ module.exports = async inputs => {
             );
         }
 
-        const output = await execute(inputs);
+        const { output, duration } = await execute(inputs);
 
         // Run app state hooks
-        const webinyJs = resolve("webiny.js");
         if (!fs.existsSync(webinyJs)) {
-            console.log(`\n🎉 Done!`);
+            console.log(
+                `⚠️ ${green("webiny.js")} was not found at ${green(
+                    webinyJs
+                )}, skipping processing of hooks.`
+            );
+            console.log(`\n🎉 Done! Deploy finished in ${green(duration + "s")}.`);
             return;
         }
 
-        const config = require(webinyJs);
         for (let i = 0; i < config.apps.length; i++) {
             const app = config.apps[i];
             const appLocation = resolve(app.location);
@@ -53,11 +112,15 @@ module.exports = async inputs => {
                     await hooks.stateChanged({ env, state: output });
                 }
             } catch (err) {
-                // if webiny.js file is not found in the app, ignore it.
+                console.log(
+                    `⚠️ ${red(err.message)}, while processing hooks at ${green(appLocation)}.`
+                );
             }
         }
 
-        console.log(`\n🎉 Done!`);
+        console.log(`\n🎉 Done! Deploy finished in ${green(duration + "s")}.`);
+        notify({ message: `API deploy completed in ${duration}s.` });
+
         if (output.cdn) {
             console.log(`🚀 GraphQL API URL: ${green(output.cdn.url + "/graphql")}`);
             if (isFirstDeploy) {