+++ scripts/jest/config.base.js
@@ -12,6 +12,6 @@ module.exports = {
     rootDir: process.cwd(),
     testRegex: `packages/(${packages.join("|")})/.*test.js$`,
     // testEnvironment: "node",
-    modulePathIgnorePatterns: ["dist", ".verdaccio", "build"],
+    modulePathIgnorePatterns: ["dist", ".verdaccio", "build", "packages/cli/create"],
     preset: "@shelf/jest-mongodb"
 };