+++ packages/app-admin/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/webiny/webiny-js/compare/@webiny/app-admin@0.3.2...@webiny/app-admin@0.3.3) (2019-10-24)
+
+**Note:** Version bump only for package @webiny/app-admin
+
+
+
+
+
 ## [0.3.2](https://github.com/webiny/webiny-js/compare/@webiny/app-admin@0.3.1...@webiny/app-admin@0.3.2) (2019-10-23)