+++ packages/app-i18n/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/webiny/webiny-js/compare/@webiny/app-i18n@0.2.2...@webiny/app-i18n@0.2.3) (2019-10-24)
+
+**Note:** Version bump only for package @webiny/app-i18n
+
+
+
+
+
 ## [0.2.2](https://github.com/webiny/webiny-js/compare/@webiny/app-i18n@0.2.1...@webiny/app-i18n@0.2.2) (2019-10-23)
 
 **Note:** Version bump only for package @webiny/app-i18n